# packages/specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.15.0
  description: |
    Spec-first HTTP API. Paths include the versioned prefix (`/api/v1`) so
    generated clients can be used without additional base-URL rewriting.

    **Authentication endpoints** live under `/api/v1/auth/*` and follow the
    [Better Auth protocol](https://www.better-auth.com/docs). They are
    deliberately not enumerated here: Better Auth owns that wire contract and
    evolves it independently of our OpenAPI surface. The `cookieAuth` and
    `bearerAuth` security schemes documented below describe how the *rest* of
    this API accepts credentials issued by that namespace — session cookies
    (web) and bearer tokens (mobile, via the `bearer` plugin).
  contact:
    name: App Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/catalog/instructors:
    get:
      operationId: listInstructors
      summary: List instructors with optional search
      description: |
        Returns a paginated list of instructors. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. Results are
        ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
      responses:
        "200":
          description: Instructor list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvins0000000000000000001
                        slug: andrei-neagoie
                        displayName: Andrei Neagoie
                        externalIds:
                          - source: udemy
                            externalId: "udemy:instructor:42"
                            url: "https://udemy.com/user/42"
                        coursesTotal: 12
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
  /api/v1/catalog/instructors/{slug}:
    get:
      operationId: getInstructor
      summary: Get a single instructor by slug
      description: Returns the instructor details plus a list of their associated courses.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-safe slug identifying the instructor.
          schema:
            $ref: "#/components/schemas/EntitySlug"
      responses:
        "200":
          description: Instructor found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorDetailDto"
              examples:
                found:
                  value:
                    instructor:
                      id: clxvins0000000000000000001
                      slug: andrei-neagoie
                      displayName: Andrei Neagoie
                      externalIds:
                        - source: udemy
                          externalId: "udemy:instructor:42"
                          url: "https://udemy.com/user/42"
                      coursesTotal: 1
                      createdAt: "2026-05-23T10:00:00Z"
                      updatedAt: "2026-05-23T10:00:00Z"
                    courses:
                      - id: clxvcrs0000000000000000001
                        libraryId: clxvp1234567890abcdefghij
                        slug: complete-web-developer-bootcamp
                        title: The Complete Web Developer in 2024
                        sections: []
                        progress:
                          percent: 0
                          lessonsCompleted: 0
                          lessonsTotal: 300
                        createdAt: "2026-05-20T08:00:00Z"
                        updatedAt: "2026-05-20T08:00:00Z"
                    coursesTotal: 1
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "404":
          description: Instructor not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: No instructor with slug 'andrei-neagoie' exists.
  /api/v1/catalog/studios:
    get:
      operationId: listStudios
      summary: List studios with optional search
      description: |
        Returns a paginated list of studios. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. Results are
        ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
      responses:
        "200":
          description: Studio list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvstu0000000000000000001
                        slug: zero-to-mastery
                        displayName: Zero To Mastery
                        externalIds:
                          - source: udemy
                            externalId: "udemy:organization:ztm"
                            url: "https://udemy.com/organization/ztm"
                        coursesTotal: 5
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
  /api/v1/catalog/studios/{slug}:
    get:
      operationId: getStudio
      summary: Get a single studio by slug
      description: Returns the studio details plus a list of their associated courses.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-safe slug identifying the studio.
          schema:
            $ref: "#/components/schemas/EntitySlug"
      responses:
        "200":
          description: Studio found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioDetailDto"
              examples:
                found:
                  value:
                    studio:
                      id: clxvstu0000000000000000001
                      slug: zero-to-mastery
                      displayName: Zero To Mastery
                      externalIds:
                        - source: udemy
                          externalId: "udemy:organization:ztm"
                          url: "https://udemy.com/organization/ztm"
                      coursesTotal: 1
                      createdAt: "2026-05-23T10:00:00Z"
                      updatedAt: "2026-05-23T10:00:00Z"
                    courses:
                      - id: clxvcrs0000000000000000001
                        libraryId: clxvp1234567890abcdefghij
                        slug: complete-web-developer-bootcamp
                        title: The Complete Web Developer in 2024
                        sections: []
                        progress:
                          percent: 0
                          lessonsCompleted: 0
                          lessonsTotal: 300
                        createdAt: "2026-05-20T08:00:00Z"
                        updatedAt: "2026-05-20T08:00:00Z"
                    coursesTotal: 1
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "404":
          description: Studio not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: No studio with slug 'zero-to-mastery' exists.
  /api/v1/catalog/tags:
    get:
      operationId: listTags
      summary: List tags with optional search and category filter
      description: |
        Returns a paginated list of tags. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. The
        optional `category` parameter filters by exact category value.
        Results are ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: Exact match filter on `category`. Omit to return tags from all categories.
          schema:
            type: string
      responses:
        "200":
          description: Tag list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvtag0000000000000000001
                        slug: javascript
                        displayName: JavaScript
                        category: language
                        externalIds: []
                        coursesTotal: 8
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
  /api/v1/catalog/tags/{slug}:
    get:
      operationId: getTag
      summary: Get a single tag by slug
      description: Returns the tag details plus a list of associated courses.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-safe slug identifying the tag.
          schema:
            $ref: "#/components/schemas/EntitySlug"
      responses:
        "200":
          description: Tag found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDetailDto"
              examples:
                found:
                  value:
                    tag:
                      id: clxvtag0000000000000000001
                      slug: javascript
                      displayName: JavaScript
                      category: language
                      externalIds: []
                      coursesTotal: 1
                      createdAt: "2026-05-23T10:00:00Z"
                      updatedAt: "2026-05-23T10:00:00Z"
                    courses:
                      - id: clxvcrs0000000000000000001
                        libraryId: clxvp1234567890abcdefghij
                        slug: complete-web-developer-bootcamp
                        title: The Complete Web Developer in 2024
                        sections: []
                        progress:
                          percent: 0
                          lessonsCompleted: 0
                          lessonsTotal: 300
                        createdAt: "2026-05-20T08:00:00Z"
                        updatedAt: "2026-05-20T08:00:00Z"
                    coursesTotal: 1
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "404":
          description: Tag not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: No tag with slug 'javascript' exists.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Better Auth bearer token
  schemas:
...
    EntitySlug:
      type: string
      pattern: "^[a-z0-9](?:[a-z0-9-]{0,98}[a-z0-9])?$"
      minLength: 1
      maxLength: 100
      description: >-
        URL-safe slug. 1–100 chars, lowercase ASCII letters, digits, and hyphens;
        cannot start or end with a hyphen. Shared by Instructor, Studio, and Tag aggregates.
      example: andrei-neagoie
    ExternalIdRef:
      type: object
      required: [source, externalId]
      description: A reference to an entity in an external system (e.g. Udemy, YouTube).
      properties:
        source:
          type: string
          minLength: 1
          description: >-
            Namespace identifying the external system (e.g. `udemy`,
            `youtube`). Scrapers are responsible for namespacing their ids
            (e.g. `youtube:playlist:PLxxx` vs `youtube:channel:UCyyy`).
        externalId:
          type: string
          minLength: 1
          description: Identifier within the source system.
        url:
          type: string
          format: uri
          description: Optional canonical URL of the entity on the source platform.
      examples:
        - source: udemy
          externalId: "udemy:instructor:42"
          url: "https://udemy.com/user/42"
        - source: youtube
          externalId: "youtube:channel:UCnUYZLuoy1rq1aVMwx4aTzw"
    InstructorDetailDto:
      type: object
      required: [instructor, courses, coursesTotal]
      description: Full instructor view including their associated courses (paginated, up to 20).
      properties:
        instructor:
          $ref: "#/components/schemas/InstructorDto"
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseDto"
          description: Courses associated with this instructor, up to 20, sorted by title.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses associated with this instructor (may exceed `courses.length`).
      example:
        instructor:
          id: clxvins0000000000000000001
          slug: andrei-neagoie
          displayName: Andrei Neagoie
          externalIds:
            - source: udemy
              externalId: "udemy:instructor:42"
              url: "https://udemy.com/user/42"
          coursesTotal: 1
          createdAt: "2026-05-23T10:00:00Z"
          updatedAt: "2026-05-23T10:00:00Z"
        courses: []
        coursesTotal: 1
    InstructorDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of an Instructor aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable instructor name.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this instructor.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this instructor.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvins0000000000000000001
        slug: andrei-neagoie
        displayName: Andrei Neagoie
        externalIds:
          - source: udemy
            externalId: "udemy:instructor:42"
            url: "https://udemy.com/user/42"
        coursesTotal: 12
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
    InstructorListDto:
      type: object
      required: [items, total, offset, limit]
      description: Paginated list of instructors.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InstructorDto"
        total:
          type: integer
          minimum: 0
          description: Total number of instructors matching the filter (before pagination).
        offset:
          type: integer
          minimum: 0
          description: Number of items skipped.
        limit:
          type: integer
          minimum: 1
          description: Maximum items returned per page.
      example:
        items:
          - id: clxvins0000000000000000001
            slug: andrei-neagoie
            displayName: Andrei Neagoie
            externalIds:
              - source: udemy
                externalId: "udemy:instructor:42"
                url: "https://udemy.com/user/42"
            coursesTotal: 12
            createdAt: "2026-05-23T10:00:00Z"
            updatedAt: "2026-05-23T10:00:00Z"
        total: 1
        offset: 0
        limit: 20
...
    StudioDetailDto:
      type: object
      required: [studio, courses, coursesTotal]
      description: Full studio view including their associated courses (paginated, up to 20).
      properties:
        studio:
          $ref: "#/components/schemas/StudioDto"
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseDto"
          description: Courses associated with this studio, up to 20, sorted by title.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses associated with this studio (may exceed `courses.length`).
      example:
        studio:
          id: clxvstu0000000000000000001
          slug: zero-to-mastery
          displayName: Zero To Mastery
          externalIds:
            - source: udemy
              externalId: "udemy:organization:ztm"
              url: "https://udemy.com/organization/ztm"
          coursesTotal: 5
          createdAt: "2026-05-23T10:00:00Z"
          updatedAt: "2026-05-23T10:00:00Z"
        courses: []
        coursesTotal: 5
    StudioDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of a Studio aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable studio name.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this studio.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this studio.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvstu0000000000000000001
        slug: zero-to-mastery
        displayName: Zero To Mastery
        externalIds:
          - source: udemy
            externalId: "udemy:organization:ztm"
            url: "https://udemy.com/organization/ztm"
        coursesTotal: 5
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
    StudioListDto:
      type: object
      required: [items, total, offset, limit]
      description: Paginated list of studios.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudioDto"
        total:
          type: integer
          minimum: 0
          description: Total number of studios matching the filter (before pagination).
        offset:
          type: integer
          minimum: 0
          description: Number of items skipped.
        limit:
          type: integer
          minimum: 1
          description: Maximum items returned per page.
      example:
        items:
          - id: clxvstu0000000000000000001
            slug: zero-to-mastery
            displayName: Zero To Mastery
            externalIds:
              - source: udemy
                externalId: "udemy:organization:ztm"
                url: "https://udemy.com/organization/ztm"
            coursesTotal: 5
            createdAt: "2026-05-23T10:00:00Z"
            updatedAt: "2026-05-23T10:00:00Z"
        total: 1
        offset: 0
        limit: 20
...
    TagDetailDto:
      type: object
      required: [tag, courses, coursesTotal]
      description: Full tag view including associated courses (paginated, up to 20).
      properties:
        tag:
          $ref: "#/components/schemas/TagDto"
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseDto"
          description: Courses associated with this tag, up to 20, sorted by title.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses associated with this tag (may exceed `courses.length`).
      example:
        tag:
          id: clxvtag0000000000000000001
          slug: javascript
          displayName: JavaScript
          category: language
          externalIds: []
          coursesTotal: 8
          createdAt: "2026-05-23T10:00:00Z"
          updatedAt: "2026-05-23T10:00:00Z"
        courses: []
        coursesTotal: 8
    TagDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of a Tag aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable tag name.
        category:
          oneOf:
            - type: string
              maxLength: 64
            - type: "null"
          description: Free-form category label (e.g. `language`, `framework`, `topic`). Null when uncategorised.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this tag.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this tag.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvtag0000000000000000001
        slug: javascript
        displayName: JavaScript
        category: language
        externalIds: []
        coursesTotal: 8
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
    TagListDto:
      type: object
      required: [items, total, offset, limit]
      description: Paginated list of tags.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TagDto"
        total:
          type: integer
          minimum: 0
          description: Total number of tags matching the filter (before pagination).
        offset:
          type: integer
          minimum: 0
          description: Number of items skipped.
        limit:
          type: integer
          minimum: 1
          description: Maximum items returned per page.
      example:
        items:
          - id: clxvtag0000000000000000001
            slug: javascript
            displayName: JavaScript
            category: language
            externalIds: []
            coursesTotal: 8
            createdAt: "2026-05-23T10:00:00Z"
            updatedAt: "2026-05-23T10:00:00Z"
        total: 1
        offset: 0
        limit: 20
...
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: ["string", "null"]
          format: uri
          default: about:blank
        title:
          type: ["string", "null"]
        status:
          type: integer
          format: int32
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        code:
          type: ["string", "null"]
...